refactor(auth): move refresh token logic to login/tokenUtils.ts as per maintainer feedback
